// src/types/post.ts
export interface ICreateSocialPost {
  audience: Audience;
  text: Descendant[];
  gifs: string;
  location: ILocationSelect | null;
  editMode: boolean;
  id?: string;
  currentPosition: number[];
  mediaUrls: MediaUrlInputType[];
  fileWithError: string;
}

export interface ICreateComments {
  text: Descendant[];
}